+++ src/url_encoded.rs
@@ -4,8 +4,9 @@ use core::fmt;
 
 use serde::de::Error;
 
-/// The error returned when attempting to decode a character in a [UrlEncodedString].
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+/// The error returned when attempting to decode a character in a [UrlEncodedString].
 pub enum UrlEncodedCharacterDecodeError {
     /// Percent symbol is not followed by two hex digits.
     BadlyFormattedPercentEncoding,
@@ -176,8 +177,9 @@ struct UrlEncodedRepresentation<'a> {
     value: &'a str,
 }
 
-/// The error returned when attempting to decode a [UrlEncodedString] into a string.
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+/// The error returned when attempting to decode a [UrlEncodedString] into a string.
 pub enum DecodeError {
     /// Error during decoding a character.
     BadUrlEncodedCharacter(UrlEncodedCharacterDecodeError),
@@ -410,6 +412,7 @@ impl<'de> serde::Deserializer<'de> for DeserializeUrlEncoded<'de> {
 }
 
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 /// Failed to deserialize a URL-Encoded form
 pub struct FormDeserializationError;